Sceneフォルダの作成、startのcssの適用、selectionのejsテンプレートの修正
