+++ public/style.css
@@ -1,0 +1,11 @@
+.start {
+    margin: 0 auto;
+    margin-top: 100px;
+    width: 80%;
+    max-width: 500px;
+    text-align: center;
+}
+
+.start p {
+    margin-top: 50px;
+}
