+++ public/templates/selection.ejs
@@ -1,4 +1,3 @@
-<h1>Hello world <img src="images/apple.jpg"/></h2>
 <h2>Select a stage</h2>
 <% for(var stage of stages) { %>
 <p id="<%= stage.id %>" >  <%= stage.name %></p>
